CBQE-1116: app.init remove queues from last run
by default remove any workload/queries queues. when --purge is used also clear up cc_queues

Reviewed-on: http://review.couchbase.org/25405
Tested-by: Tommie McAfee <tommie@couchbase.com>

+++ pysystests/app/init.py
@@ -1,4 +1,5 @@
 import os
+import sys
 from rabbit_helper import RabbitHelper
 from cache import CacheHelper
 import testcfg as cfg
@@ -9,7 +10,6 @@ os.system("mkdir -p "+cfg.LOGDIR)
 #make sure celeybeat-schedule.db file is deleted
 os.system("rm -rf celerybeat-schedule.db")
 
-CacheHelper.cacheClean()
 
 # kill old background processes
 kill_procs=["sdkserver"]
@@ -45,17 +45,37 @@ if cfg.CLUSTER:
 
         os.system("cd .. && ./testrunner -i " + cfg.CLUSTER['ini'] + " -t cluster_setup.SETUP." + function + " && cd pysystests")
 
-# cleaning up seriesly database (fast and slow created by cbtop)
+# delete queues (note using --purge will remove cc_queues)
+queues = CacheHelper.task_queues()
 
-if cfg.SERIESLY_IP != '':
-    from seriesly import Seriesly
-    os.system("curl -X DELETE http://{0}:3133/fast".format(cfg.SERIESLY_IP))
-    os.system("curl -X DELETE http://{0}:3133/slow".format(cfg.SERIESLY_IP))
-    os.system("curl -X DELETE http://{0}:3133/event".format(cfg.SERIESLY_IP))
-    os.system("curl -X DELETE http://{0}:3133/atop".format(cfg.SERIESLY_IP))
+# when --purge set delete cc_queue's as well
+# as seriesly db
+if "--purge" in sys.argv:
 
-    seriesly = Seriesly(cfg.SERIESLY_IP, 3133)
-    seriesly.create_db('event')
+    queues = set(CacheHelper.queues())
+
+    # cleaning up seriesly database (fast and slow created by cbtop)
+    if cfg.SERIESLY_IP != '':
+        from seriesly import Seriesly
+        os.system("curl -X DELETE http://{0}:3133/fast".format(cfg.SERIESLY_IP))
+        os.system("curl -X DELETE http://{0}:3133/slow".format(cfg.SERIESLY_IP))
+        os.system("curl -X DELETE http://{0}:3133/event".format(cfg.SERIESLY_IP))
+        os.system("curl -X DELETE http://{0}:3133/atop".format(cfg.SERIESLY_IP))
+
+        seriesly = Seriesly(cfg.SERIESLY_IP, 3133)
+        seriesly.create_db('event')
+
+
+
+for q_ in queues:
+    try:
+        RabbitHelper().delete(q_)
+        print "Cleanup Queue: %s" % q_
+    except Exception as ex:
+        pass
+
+# clean up cache
+CacheHelper.cacheClean()
 
 # start sdk server
 os.system("python sdkserver.py  &")
